MF-502: allow submits
