+++ src/patient-registration/before-save-prompt.tsx
@@ -1,32 +1,25 @@
-import React, { useCallback, useState, useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
+import React, { useEffect } from 'react';
 import Modal from 'carbon-components-react/es/components/Modal';
 import { useTranslation } from 'react-i18next';
 
-const getUrlWithoutPrefix = url => url.split(window['getOpenmrsSpaBase']())?.[1];
 const noop = () => {};
 
 interface BeforeSavePromptProps {
   when: boolean;
+  open: boolean;
+  cancelNavFn: EventListener;
+  onRequestClose: Function;
+  onRequestSubmit: Function;
 }
 
-const BeforeSavePrompt: React.FC<BeforeSavePromptProps> = ({ when }) => {
-  const history = useHistory();
-  const [open, setOpen] = useState(false);
-  const [newUrl, setNewUrl] = useState(undefined);
+const BeforeSavePrompt: React.FC<BeforeSavePromptProps> = ({
+  when,
+  open,
+  cancelNavFn,
+  onRequestClose,
+  onRequestSubmit,
+}) => {
   const { t } = useTranslation();
-
-  const cancelNavFn = useCallback(evt => {
-    if (!open && !evt.detail.navigationIsCanceled) {
-      evt.detail.cancelNavigation();
-      setNewUrl(evt.detail.newUrl);
-      setOpen(true);
-
-      // once the listener is run, we want to remove it immediately in case an infinite loop occurs due
-      // to constant redirects
-      evt.target.removeEventListener('single-spa:before-routing-event', cancelNavFn);
-    }
-  }, []);
 
   useEffect(() => {
     if (when) {
@@ -38,8 +31,9 @@ const BeforeSavePrompt: React.FC<BeforeSavePromptProps> = ({ when }) => {
 
     return () => {
       window.onbeforeunload = noop;
+      window.removeEventListener('single-spa:before-routing-event', cancelNavFn);
     };
-  }, [when]);
+  }, [when, cancelNavFn]);
 
   return (
     <Modal
@@ -49,14 +43,8 @@ const BeforeSavePrompt: React.FC<BeforeSavePromptProps> = ({ when }) => {
         modalHeading: t('discardModalHeader'),
         primaryButtonText: t('discard'),
         secondaryButtonText: t('cancel'),
-        onRequestClose: () => {
-          setOpen(false);
-          // add the route blocked when
-          window.addEventListener('single-spa:before-routing-event', cancelNavFn);
-        },
-        onRequestSubmit: () => {
-          history.push(`/${getUrlWithoutPrefix(newUrl)}`);
-        },
+        onRequestClose,
+        onRequestSubmit,
       }}>
       {t('discardModalBody')}
     </Modal>
